apps/services/cipas-service/pyproject.toml
[tool.poetry]
name = "cipas-service"
version = "0.1.0"
description = "CIPAS — Code Integrity Analysis Service (async-first FastAPI microservice)"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "cipas", from = "src" }
]
keywords = ["fastapi", "async", "minio", "redis", "observability", "prometheus", "opentelemetry"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# Web framework + ASGI server
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}

# Settings / config
pydantic-settings = "^2.2.0"

# Observability & logging
loguru = "^0.7.0"
prometheus-fastapi-instrumentator = "^6.0.0"
opentelemetry-api = "^1.17.0"

# HTTP client (async)
httpx = "^0.24.0"

# Utilities
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = {version = "^0.24.0", extras = ["http2"]}
ruff = "^0.12.0"
mypy = "^1.5.0"
black = "^24.3.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

# Tooling configs -------------------------------------------------------------

[tool.ruff]
line-length = 88
target-version = "py311"
extend-select = ["E", "F"]
# Enable common strict rules; tune as needed
ignore = ["E501"]
exclude = ["src/cipas/__pycache__", "build", ".venv", "dist", "tests"]

[tool.mypy]
python_version = 3.11
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
plugins = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q -ra"
testpaths = ["tests"]

# Optional convenience scripts for local development
[tool.poetry.scripts]
cipas = "cipas.main:app"

# Recommendations / metadata for maintainers
[tool.metadata]
maintainers = ["platform-team@example.com"]
repository = "https://example.com/your-org/gradeloop-core-v2"

# End of file
